Move libs from lib/inc/header.h to lib/inc/lib/header.h so in lib2 they are referenced as #include <lib/header.h> not <header.h>
