+++ Delta++/bench/benchmark.cpp
@@ -1,6 +1,6 @@
 #include <benchmark/benchmark.h>
 
-#include "engine_factory.h"
+#include <Delta++/engine_factory.h>
 using namespace DPP;
 using namespace std::string_literals;
