+++ Delta++/inc/Delta++/abstract_engine.h
