+++ Delta++/inc/Delta++/calc.h
