+++ Delta++/inc/Delta++/trade.h
