+++ Delta++/inc/Delta++/triangular_matrix.h
